refactor(app): extract useSessionInteractions from session-route
Permissions/questions/todos cluster (~170 lines: busy state, query-cache
subscriptions, snapshot seed effects, reply handlers) moves verbatim to
domains/session/sync/use-session-interactions.ts. Nothing in the component
read this cluster's internals except the JSX, so the seam is a single hook
call with four stable inputs (client, workspaceId, sessionId, workspaceRoot).

The generic useQueryCacheState helper moves to infra/query-cache-state.ts
next to the query-client singleton.

session-route.tsx: 3,161 -> 3,020 lines (3,328 at phase start).

Tested:
- pnpm typecheck clean; bun test tests/ 58 pass
- pnpm build clean
- pnpm test:e2e — 23/23 assertions ok (boots real app, exercises sessions,
  permissions among them)
